+++ assets/scripts/ui/kaishiyouxi.ts
@@ -1,6 +1,6 @@
 
 import { _decorator, Component, director } from 'cc';
-import { AudioManager } from '../common/AudioManager';
+import { AudioManager } from '../common/AudioManager.ts';
 const { ccclass, property } = _decorator;
 
 @ccclass('kaishiyouxi')
